Move public classes to new files
