+++ src/Microsoft.ML.OnnxTransformer/OnnxTypeParser.cs
@@ -14,7 +14,7 @@ using Microsoft.ML.Runtime;
 
 namespace Microsoft.ML.Transforms.Onnx
 {
-    internal class OnnxTypeHelper
+    internal static class OnnxTypeParser
     {
         /// <summary>
         /// Derive the corresponding <see cref="Type"/> for ONNX tensor's element type specified by <paramref name="dataType"/>.
@@ -363,185 +363,4 @@ namespace Microsoft.ML.Transforms.Onnx
         }
     }
 
-    /// <summary>
-    /// The corresponding <see cref="DataViewSchema.Column.Type"/> of ONNX's sequence type in <see cref="IDataView"/>'s type system.
-    /// In other words, if an ONNX model produces a sequence, a column in <see cref="IDataView"/> may be typed to <see cref="OnnxSequenceType"/>.
-    /// Its underlying type is <see cref="IEnumerable{T}"/>, where the generic type "T" is the input argument of
-    /// <see cref="OnnxSequenceType.OnnxSequenceType(Type)"/>.
-    /// </summary>
-    public sealed class OnnxSequenceType : StructuredDataViewType
-    {
-        private static Type MakeNativeType(Type elementType)
-        {
-            var enumerableTypeInfo = typeof(IEnumerable<>);
-            var enumerableType = enumerableTypeInfo.MakeGenericType(elementType);
-            return enumerableType;
-        }
-
-        /// <summary>
-        /// Create the corresponding <see cref="DataViewType"/> for ONNX sequence.
-        /// </summary>
-        /// <param name="elementType">The element type of a sequence.</param>
-        public OnnxSequenceType(Type elementType) : base(MakeNativeType(elementType))
-        {
-            DataViewTypeManager.Register(this, RawType, new[] { new OnnxSequenceTypeAttribute(elementType) });
-        }
-
-        public override bool Equals(DataViewType other)
-        {
-            if (other is OnnxSequenceType)
-                return RawType == other.RawType;
-            else
-                return false;
-        }
-
-        public override int GetHashCode()
-        {
-            return RawType.GetHashCode();
-        }
-    }
-
-    /// <summary>
-    /// The corresponding <see cref="DataViewSchema.Column.Type"/> of ONNX's map type in <see cref="IDataView"/>'s type system.
-    /// In other words, if an ONNX model produces a map, a column in <see cref="IDataView"/> may be typed to <see cref="OnnxMapType"/>.
-    /// Its underlying type is <see cref="IDictionary{TKey, TValue}"/>, where the generic type "TKey" and "TValue" are the input arguments of
-    /// <see cref="OnnxMapType.OnnxMapType(Type,Type)"/>.
-    /// </summary>
-    public sealed class OnnxMapType : StructuredDataViewType
-    {
-        /// <summary>
-        /// Create the corresponding <see cref="DataViewType"/> for ONNX map.
-        /// </summary>
-        /// <param name="keyType">Key type of the associated ONNX map.</param>
-        /// <param name="valueType">Value type of the associated ONNX map.</param>
-        public OnnxMapType(Type keyType, Type valueType) : base(typeof(IDictionary<,>).MakeGenericType(keyType, valueType))
-        {
-            DataViewTypeManager.Register(this, RawType, new[] { new OnnxMapTypeAttribute(keyType, valueType) });
-        }
-
-        public override bool Equals(DataViewType other)
-        {
-            if (other is OnnxMapType)
-                return RawType == other.RawType;
-            else
-                return false;
-        }
-
-        public override int GetHashCode()
-        {
-            return RawType.GetHashCode();
-        }
-    }
-
-    /// <summary>
-    /// To declare <see cref="OnnxSequenceType"/> column in <see cref="IDataView"/> as a field
-    /// in a <see langword="class"/>, the associated field should be marked with <see cref="OnnxSequenceTypeAttribute"/>.
-    /// Its uses are similar to those of <see cref="VectorTypeAttribute"/> and other <see langword="class"/>es derived
-    /// from <see cref="DataViewTypeAttribute"/>.
-    /// </summary>
-    public sealed class OnnxSequenceTypeAttribute : DataViewTypeAttribute
-    {
-        private Type _elemType;
-
-        /// <summary>
-        /// Create a sequence type.
-        /// </summary>
-        public OnnxSequenceTypeAttribute()
-        {
-        }
-
-        /// <summary>
-        /// Create a <paramref name="elemType"/>-sequence type.
-        /// </summary>
-        public OnnxSequenceTypeAttribute(Type elemType)
-        {
-            _elemType = elemType;
-        }
-
-        /// <summary>
-        /// Sequence types with the same element type should be equal.
-        /// </summary>
-        public override bool Equals(DataViewTypeAttribute other)
-        {
-            if (other is OnnxSequenceTypeAttribute otherSequence)
-                return _elemType.Equals(otherSequence._elemType);
-            return false;
-        }
-
-        /// <summary>
-        /// Produce the same hash code for sequence types with the same element type.
-        /// </summary>
-        public override int GetHashCode()
-        {
-            return _elemType.GetHashCode();
-        }
-
-        /// <summary>
-        /// An implementation of <see cref="DataViewTypeAttribute.Register"/>.
-        /// </summary>
-        public override void Register()
-        {
-            var enumerableType = typeof(IEnumerable<>);
-            var type = enumerableType.MakeGenericType(_elemType);
-            DataViewTypeManager.Register(new OnnxSequenceType(_elemType), type, new[] { this });
-        }
-    }
-
-    /// <summary>
-    /// To declare <see cref="OnnxMapType"/> column in <see cref="IDataView"/> as a field
-    /// in a <see langword="class"/>, the associated field should be marked with <see cref="OnnxMapTypeAttribute"/>.
-    /// Its uses are similar to those of <see cref="VectorTypeAttribute"/> and other <see langword="class"/>es derived
-    /// from <see cref="DataViewTypeAttribute"/>.
-    /// </summary>
-    public sealed class OnnxMapTypeAttribute : DataViewTypeAttribute
-    {
-        private Type _keyType;
-        private Type _valueType;
-
-        /// <summary>
-        /// Create a map (aka dictionary) type.
-        /// </summary>
-        public OnnxMapTypeAttribute()
-        {
-        }
-
-        /// <summary>
-        /// Create a map (aka dictionary) type. A map is a collection of key-value
-        /// pairs. <paramref name="keyType"/> specifies the type of keys and <paramref name="valueType"/>
-        /// is the type of values.
-        /// </summary>
-        public OnnxMapTypeAttribute(Type keyType, Type valueType)
-        {
-            _keyType = keyType;
-            _valueType = valueType;
-        }
-
-        /// <summary>
-        /// Map types with the same key type and the same value type should be equal.
-        /// </summary>
-        public override bool Equals(DataViewTypeAttribute other)
-        {
-            if (other is OnnxMapTypeAttribute otherSequence)
-                return _keyType.Equals(otherSequence._keyType) && _valueType.Equals(otherSequence._valueType);
-            return false;
-        }
-
-        /// <summary>
-        /// Produce the same hash code for map types with the same key type and the same value type.
-        /// </summary>
-        public override int GetHashCode()
-        {
-            return Hashing.CombineHash(_keyType.GetHashCode(), _valueType.GetHashCode());
-        }
-
-        /// <summary>
-        /// An implementation of <see cref="DataViewTypeAttribute.Register"/>.
-        /// </summary>
-        public override void Register()
-        {
-            var enumerableType = typeof(IDictionary<,>);
-            var type = enumerableType.MakeGenericType(_keyType, _valueType);
-            DataViewTypeManager.Register(new OnnxMapType(_keyType, _valueType), type, new[] { this });
-        }
-    }
 }
